Refactor(performance): Split the git optimizer
git_optimizer.py held 708 lines, past the 400-line reviewability limit
the aislop code-quality gate applies.

Split it into a package at the same dotted path: models (the strategy
and operation enums plus the config/result records), strategies
(ShallowCloneStrategy and ReferenceRepository), optimizer (the
GitOptimizer coordinator) and utils (the two module-level helpers the
original already fenced off under a banner comment). The package
__init__ re-exports the previous names.
